Fix incorrect J9Class sign extension
When throwing a ClassCastException from jit'd code on x86 64bit
compressed refs it is possible for a J9Class to be incorrectly
sign extended, causing a segfault. This change avoids the problem
by detecting if the (constant) class would be affected and
using the existing code for 64 bit pointers instead of the
problematic 32 bit code.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>