Store caches inside ClassInfo instead of pointers to them
[skip ci]
Caches that are stored inside ClassInfo are global and persist
until the class is unloaded or client disconnects. Thus, most
caches will likely be populated, so it's more convinient to store the
cache outright, instead of a pointer to the cache. This removes the
need for allocating memory for the cache the first time it's used and
freeing it when class info is freed.
Even if the cache is never used, empty `std::unordered_map` takes little
space so it shouldn't have a significant effect on memory footprint.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>