Explicitly check if ROM class is in SCC
Before requesting the SCC offset of a ROM class we need to first check
that it's in the SCC. This is because offsetInSharedCacheFromPointer()
no longer blindly calculates offsets for whatever is passed in, it now
checks that the parameter is in the SCC before calculating its offset.

isPointerInSharedCache() is the more appropriate API, it returns a bool
indicating whether the parameter is in the SCC and if so returns its
offset (if the caller requested one).

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>