Order relationship list by classname lengths
Currently, the code always adds the relationship even if it has already
been added.  This can lead to very long lists of parents which delays
processing them.

The insight here is that by keeping the list in the order of the
classname length, we can more quickly weed out duplicates as we only
need to walk the list until there is a longer name, or the same name
already present, before inserting.

Additionally, avoid the parent node allocation until we know we’re
adding something to the list.

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>