Validate that stored class relationships are correct
When a class is being initialized, validate that it holds the expected
relationship with each parent in its list.

Free memory for the child hash table entry and parent nodes once
validation is complete.

Throw java/lang/VerifyError if validation of class relationships fails.

Note that this validation check can be moved to any point prior to
a class being made available.

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>