Handle Throwable parent class relationship as special case
Instead of allocating a parent node for Throwable, set a bit in the
childEntry flags J9RELATIONSHIP_PARENT_IS_THROWABLE.

During validation, check if the bit is set. If so, retrieve the Throwable
class and check that it is a superclass of the child.

Throwable seems to occur frequently as a parent class, so this change
will avoid the duplicate memory allocations done for Throwable.

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>