Cache which fields are unresolved in TR_J9ServerVM::jitFieldsAreSame
[skip ci]
`VM_jitFieldsAreSame` is the most frequent message in all
of the benchmarks we use.
We already perform global caching of resolved fields, but we cannot do
that for unresolved fields because they might become resolved later on.
A workaround that this commit implements is to cache which fields are
unresolved
for the duration of a single compilation. This way, even if a field
becomes resolved during the compilation, incorrect result will be used
only for a short period of time, not affecting performance.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>