Remove assert in Monitor Table Read Barrier
The assert served its purpose - after months of usage, it 'proved' that
copy is indeed not required in weak root access barrier.

But the assert may occasionally trigger and introduce confusion, if it
is invoked at spots where object liveness is unknown (but still, we it's
safe to access the object without coping it, since noone is racing to
mutate the object). An example is JLM iterating over monitor table
entries, under exclusive VM access, to dump info for all monitors. To
avoid confusion and false asserts, we are removing it.

See also:
https://github.com/eclipse/openj9/pull/5232

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>