Add tests for -XX:[+|-]ClassRelationshipVerifier
Test the following cases:
- only source class loaded
- only target class loaded
- neither target nor source classes loaded
- source is subclass of target
- source implements target
- source is not subclass of target
- source does not implement target

For tests where classes are not already loaded, query the
classRelationships hash table in the class loader to check that a
source/child and target/parent relationship has been recorded.

TestClassRelationshipVerifier is excluded from testvmcheck since
it is incompatible with that test and does not benefit from being
tested with `-Xverify:all` or `-Xcheck:vm`.

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>