Use J9RecognizedCallTransformer for rotateLeft on Power
Previously, the Integer.rotateLeft and Long.rotateLeft recognized
methods were being handled in the codegen on Power. This is completely
unnecessary, since the Power codegen supports the irol and lrol opcodes
that J9RecognizedCallTransformer uses for these methods. It is also
suboptimal, since it prevents the optimizer from performing certain
optimizations on these calls due to the fact that it cannot see through
the call. To bring the Power codegen into alignment with the other
codegens, handling of these methods has been moved to
J9RecognizedCallTransformer on Power.

Signed-off-by: Ben Thomas <ben@benthomas.ca>