Remove J9PPCArrayCmp.spp
Previously, the Power codegen had a file called J9PPCArrayCmp.spp that
was supposed to have helpers related to array comparison. However, this
file has been surrounded by an `#if 0` guard for a significant period of
time, rendering it useless. In line with this, the file has been
entirely removed as it's not doing anything.

Signed-off-by: Ben Thomas <ben@benthomas.ca>