Set custom SecurityManager constructor accessible flag
The constructor for custom SecurityManager must set its accessible flag
to true to suppress access checking, otherwise an IllegalAccessException
is thrown when invoking constructor.newInstance().
Changed getConstructor() to getDeclaredConstructor() to only allow
instantiation of custom SecurityManger with a public constructor just
like RI.
Added a test to ensure that NoSuchMethodException is thrown at
getConstructor() when loading a custom SecurityManger with a non-public
constructor.

Signed-off-by: Jason Feng <fengj@ca.ibm.com>