Skip Pre- and PostEscapeAnalysis if EA is not enabled
The PreEscapeAnalysis and PostEscapeAnalysis optimization passes are
designed so that PreEscapeAnalysis inserts extra information into the
trees for the use of EscapeAnalysis (EA), and then PostEscapeAnalysis
undoes the modifications of PreEscapeAnalysis.  If EA does not run, the
effect of running PreEscapeAnalysis and PostEscapeAnalysis is a no-op.

This change skips performing PreEscapeAnalysis and PostEscapeAnalysis if
EA itself is not enabled.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>