Remove invalid intrinsicIndexOf* test
The invalid test breaks the contract of the API which states an
implicit assumption of the API is that the offset is at most equal to
the length passed. For the case we removed this was not true and the
JIT ends up inlining an intrinsic which uses such an assumption to
perform the intrinsicIndexOf* operation.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>