Eliminate VM_isSameOrSuperClass messages
[skip ci]
After a recent rebase `VM_isSameOrSuperClass` is the most
frequent message in AOT mode, as described in issue #7072.
Eliminate this message by using `getSuperClass` method of
the VM to fetch the super class, instead of calling a VM
method directly. Since we already do caching in `getSuperClass`,
this will eliminate most messages associated with
`isSameOrSuperClass`.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>