Port TR_J9ServerVM/TR_J9SharedCacheServerVM to master
Introduce `TR_J9ServerVM` and `TR_J9SharedCacheServerVM`. `TR_J9ServerVM` is an extension to `TR_J9VM` which overrides a large number of its APIs. The class is used by JITServer and the overridden APIs mostly send remote messages to JITClient to query information from the `TR_J9VM` on the client. The information is needed for JITServer to compile code that is compatible with the client-side VM. To minimize the number of remote messages as a way to optimize JITServer performance, a lot of client-side TR_J9VM information are cached on JITServer. `TR_J9SharedCacheServerVM` is an extension to `TR_J9ServerVM` that does similar things compared to `TR_J9ServerVM` except it's used for AOT compilations.

Signed-off-by: Harry Yu <harryyu1994@gmail.com>