Eliminate ClassEnv_getROMClassRefName messages
[skip ci]
This message is one of the most frequent messages
during AOT compilations.
The method from which it's called accesses ROM constant pool,
gets `J9ROMClassRef` from there and retrieves the name of the
reference. We already cache ROM classes and their constant pools
on the server, however, we can't simply access reference name,
because the actual class reference might be outside the ROM class,
and thus not necessarily cached. We already encountered and solved
a similar problem in resolved methods, so I just copied methods
`getROMString` and `getRemoteROMString` from there into `ClassInfo`
and used them to appropriately fetch a string from either cache or
the client.
In a follow-up commit, the resolved method versions of methods that
retrieve ROM strings should be removed and replaced with `ClassInfo`
versions.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>