Use system linkage for jprofiler c helpers on X86
Jprofiler helpers are c functions which are implemented with complex c
code, calling them on java stack can be very dangerous. One known issue
is that java stack is not 16-byte aligned on X86-32, this will cause
crashes if there's SSE instruction in the c code. Use system linkage to
switch to c stack before the call.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>