Merge PR4032 from JITServer to master
- This change merges PR #4032 that increases the number of compilation
threads at the server to 64 and it affects the implementation of
the non-JITServer mode as well.
- The maximum of 8 compilation threads limitation is due to
the size of uint8_t TR_PersistentClassInfo::_shouldNotBeNewlyExtended.
The JITaaS server uses a different mechanism to track newly extended
classes and it does not set the bit in _shouldNotBeNewlyExtended.
- For the JITClient and the non-JITServer mode, the max number
of compilation threads remains as 8.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>