Reduce VM_getClassNameSignatureFromMethod messages
This message is emitted from `TR_J9ServerVM::sampleSignature`
It requires a class name, method name and method signature.
All of these things are already cached and can be accesed
from client data. In cases when the corresponding class/method
is not cached, still need to make a remote call.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>