Fix decompile at exception catch
At an exception catch point in jitted code, there could be any number of
inlines, e.g.
```
void blah() throws Exception {
   throw new Exception();
}

void bar(Exception e) {
   e.printStackTrace();
}

void foo(Exception e) {
   bar(e);
}

void example() {
   try {
      blah();
   } catch(Exception e) {
      foo(e);
   }
}
```
The previous code was thinking like the interpreter - the catching stack
must be a subset of the throwing stack, which is not true in the example
above.

With full inlining, the stack at the catch point could be:
```
bar
foo
example
```
When decompiling at a catch, rather than depending on subsetting,
discard any existing decompilation for the frame and perform a new OSR
at the catch point with the correct stack.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>