Add wrapper around SCC API to prevent assert
Under HCR, when an agent redefines a class, it causes the J9Class
pointer to stay the same, but the J9ROMClass pointer changes. This means
that if the compiler has a reference to a J9Class whose J9ROMClass was
in the SCC at one point in the compilation, it may no longer be so at
another point in the compilation. This means that the compilation is no
longer valid and should be aborted. The way a compilation thread knows
it should abort is by checking a flag. This is only done at certain
locations. However, even if there isn't an abort during the
compilation, at the end of the  compilation the compiler will fail the
compile if such a redefinition occurred.

TR_J9SharedCache::offsetInSharedCacheFromPointer asserts if the pointer
passed in does not exist in the SCC; therefore, calling it after such a
redefinition could result in an assert.

This commit adds a wrapper around
TR_J9SharedCache::isPointerInSharedCache; isPointerInSharedCache doesn't
assert and, conveniently, updates the location referred to by the
cacheOffset pointer passed in as a parameter.

If the pointer whose offset is required is not in the SCC, the
compilation is aborted. Otherwise, the offset is used. Even if a
redefintion occurs (which would result in an incorrect offset), as
stated above, the compilation will eventually be aborted.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>