Replace binary length peeking in PPCRecompilationSnippet
Previously, there was code in PPCRecompilationSnippet which would peek
at the binary encoded length of a conditional branch to set a flag. This
code was trying to check whether the conditional branch to the snippet
was expanded into two instructions due to the branch distance being too
large for a conditional branch instruction.

While this worked in the past, the binary encoder in the Power codegen
is undergoing some work that will result in this check no longer working
correctly in the near future. Instead, PPCRecompilationSnippet will now
use getFarRelocation() to check the flag which is explicitly set when
such expansion occurs.

Signed-off-by: Ben Thomas <ben@benthomas.ca>