Delete obsolete Dockerfile's and build instructions for jdk9, 10, 12
These older, out of support versions can no longer be built with OpenJ9.

Update the Dockerfile and README in the docker/test directory to refer
to jdk11 instead of jdk9.

I didn't delete the jdk9 arm 32-bit Dockerfile as it may contain some
useful information if support ever picks up for this processor.

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>