Improved code generation for compare and branch instructions
Code generation for all compare and branch instructions has been
consolidated into `generateS390CompareAndBranchInstruction()` API.
Instructions consolidated in this commit are `CRJ` and `CIJ`.

This is allows for compare and branch instruction optimizations to be
applied through a common point. In addition to that, all compare and
branch instructions can now be to be disabled through
`Xjit:disableCompareAndBranch`.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>