Fix incorrect use of cmpli4 as PPCTrg1Src2Instruction
Due to a typo, genHeapAlloc was attempting to use the cmpli4 instruction
as a PPCTrg1Src2Instruction. This results in the binary encoding being
botched and a garbage value being encoded in the UI field of the
instruction. This has been switched to generate a cmpl4 instruction
instead, which matches the obviously intended semantics of comparing two
registers.

Signed-off-by: Ben Thomas <ben@benthomas.ca>