Fix compiler issues with SCC hint pointers
Certain versions of XL C/C++ generate incorrect
code when accessing the individual fields of a
uint32_t SCC hint stored in a local variable via
uint16_t pointers.

As a result of this behaviour we sometimes write
corrupt hints to the SCC (specifically the previous
content of the stack location the hint variable is
mapped to), which can cause performance problems.

The problem was thought to be fixed by

cda6447d9d9698f1a1c531f979e8a5f8a2077634

but that turned out to be insufficient.

This patch introduces a struct with named uint16_t
fields to represent the hint and simplifies some
of the associated code. The struct is returned by
value and because of its size, the types of its
members, its trivial copy-ctor and dtor, etc. it
should be returned in a single register on many
ABIs, making it equivalent to the old code in most
cases.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>