Always call methodCannotBeRecompiled() when failing a recompilation
Previous code had a bug where the call to methodCannotBeRecompiled()
in compilationEnd() was conditioned by the existence of a compilation
entry. However, at shutdown we want to fail all incoming compilation
requests even before they start and when doing so we don't provide a
comp object or a valid entry to compilationEnd(). Thus, we should call
methodCannotBeRecompiled() regardless of the existentence of a valid
entry.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>