Avoid load rewrites for cold escapes of immutable discontiguous objects
Ordinarily Escape Analysis avoids fix-up of fields of stack allocated
objects in cold block escapes, but as a special case, fix-up still
happens for immutable discontiguos objects, as the initialization of the
temporary used to represent a field of such an object might end up in
such a cold block.

However, the fix-up for the fields of such an object in a cold block
escape was happening for both loads and stores.  That's not usually a
problem except if the load refers to the vft-symbol, as no temporary
is ever properly initialized to represent the vft-symbol for a
discontiguously stack allocated object.

Avoided this situation by only performing the special case fix-up for
immutable discontiguously stack allocated objects for store operations.
It's safe to load the field from the heapified object.

Also added a check for the environment variable
TR_disableEAImmutableObjectHandling to allow special handle of
immutable objects by EA to be disabled.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>