Retrieve method signature using getROMString
[skip ci]
When setSignature() is called at the JITServer in
TR_ResolvedJ9JITServerMethod::getResolvedPossiblyPrivateVirtualMethod(),
the signature might not have been stored in romMethodRef at the server.
It needs to call getROMString() to retrieve the signature.
Otherwise, jitParseSignature() will be triggered with an empty signature.
It doesn’t handle the case of signature length being 0 and ends up going
into an infinite loop and causing stack overflow.

Issue: #7567

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>