Remove self() in Z linkages
Removing self() in all the Z linkages as they are not required and help
in code clarity. These changes reflect the conversation found in #7453.

The changes pertain to:
- Following files: J9Linkage.cpp / J9S390CHelperLinkage.cpp/
J9S390PrivateLinkage.cpp / J9SystemLinkagezOS.cpp .
- 0xdaryl (issue creator) explanation: Subclasses that derive from
TR::Linkage are not components of the extensible part of the hierarchy,
and they do not need to use self() methods to refer to the member
functions that are part of the extensible part of the hierarchy.
- The self() calls should be removed to make the code cleaner and to
not confuse developers into thinking these are extensible classes.
- Sanity zlinux jdk11 failed because of caching of cg() and comp() in
J9S390PrivateLinkage.cpp file.
- Build failure resolved by commenting out Caching code in above file.

Closes: #7453
Signed-off-by: Ajay Shanmuganathan <ajaypal95@gmail.com>