Prevent infinite loop in illegal field modification hook
The loop run by the JIT when notified of an illegal field write will enter an
infinite loop if we find a field to patch - we fail to update the next pointer.
This change fixes the handling of the next pointer and removes the unused prev
pointer to save code complexity since maintaining prev with the RAT mark then
sweeep scheme will be tricky and the value is not currently used.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>