Make DFPCompareTo unconditionally use mfocrf
Previously, inlining of BigDecimal.DFPCompareTo was trying to use the
mfcr instruction with an immediate operand. In the past, converting this
to an mfocrf instruction would happen during binary encoding, contingent
upon support for this instruction. However, support for performing this
transformation during binary encoding will be removed shortly as part of
a series of refactors.

In order for this code to keep working, it will need to generate an
mfocrf instruction rather than an mfcr instruction if it wishes to use
that form. This change has now been made unconditionally, since mfocrf
is supported on POWER5 and above, which includes all of our supported
configurations.

Signed-off-by: Ben Thomas <ben@benthomas.ca>