Fix JITServer message type mismatch
The main cause of the message type mismatches is due to that the JITServer/JITClient messages not being 1-to-1. In order to make sure that the messages get received by the correct receiver, we must avoid double sending of messages.

In this problem, JITClient sends a compilationRequest but immediately realized that it needs to abort the compilation so it sends another "writeError()" message. This has caused problems for both JITServer and JITClient not receiving correct messages afterwards. We resolved the issue by having the JITClient wait for the next message from JITServer before sending out the "writeError()" message.
[skip ci]

Signed-off-by: Harry Yu <harryyu1994@gmail.com>