Fix JITServer message type mismatch
The main cause of the message type mismatches
is due to JITServer/JITClient messages not
being 1-to-1. In order to make sure that the
messages always get received by the correct
receiver, we must avoid double sending of messages.

In this problem, JITClient sends a compilationRequest
but immediately realized that it needs to abort the
compilation so it sends another "writeError()"
message. This double sending from the JITClient
has caused both JITServer and JITClient not receiving
correct messages sometimes. We eliminated the double
sending and resolved the issue by having the JITClient
wait for the next message from JITServer before
sending out the "writeError()" message.

Signed-off-by: Harry Yu <harryyu1994@gmail.com>