Populate cache corruption information for non-zero layer cache
We read the cache metadata to detect whether a cache is corrupted and
populate such info to SH_OSCache_Info.isCorrupt. We do this only when
iterating the caches in a directory (reason==SHR_STATS_REASON_ITERATE).
Currently this is working on layer 0 cache only. 

This change extends this to non-zero layer cache. 
This change should only take effect when getting SH_OSCache_Info on a
compatible multi-layer cache when the reason to get cache statistics is
SHR_STATS_REASON_ITERATE.

issue #5480

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>