Fast-path ASCII strings in j9gc_createJavaLangString
Improve throughput performance of this API by introducing a forward
pass which identifies whether the entire UTF8 string is composed of
ASCII characters.

The idea is that most of the strings coming through this API are
indeed ASCII strings, typically strings relating to class, package,
and file names, in addition to some strings from the JCL. Profiling
tells us almost all of these strings are proper ASCII strings, and
as such we can take advantage of faster algorithms for translating
between UTF8 and UTF16.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>