Split entry for recognized System.arraycopy methods
The handling of System.arraycopy as a recognized method currently uses a
wildcard for its signature.  As the System.arraycopy method that takes
Object arguments is native while the method that takes array arguments
is not, separate them into two entries.  That way one can be marked as
expected to be native in a subsequent change.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>