Add missing JITServer StreamInterrupted handle
When JITServer is at the "early translation" stage
and JITClient issues an interruption, an exception
will be thrown in the next JITServer read. However,
the catch statement did not include handling of
StreamInterrupted and as a result will treat the
interruption as a "compilationFailure" though it
really should have been an interruption. An client-issued
interruption and an actual compilationFailure are
handled differently. For the actual compilationFailure,
JITserver has to tell JITclient about it. But for
an interruption, since it's issued by JITclient,
JITserver should not tell JITclient about it again.
[skip ci]

Signed-off-by: Harry Yu <harryyu1994@gmail.com>