Optimize the object allocation in stack traces
Stack traces create a string for the class name.  This patch
attempts to reuse the string that has already been allocated
in the class->classNameString field.  If the string hasn't
been allocated, we allocate an interned tenure string and
write it back to the class->classNameString field effectively
caching it for future access.

This may result in a footprint growth due to more strings being
held by the class.

This patch uses the new peekClassHashTable() api to get the
class without needing to change the iterateStackFrames call.

We also attempt to optimize the fileName string allocation by
reusing the previously allocated string in the case where
adjacent stack frames are from the same file.

Note, there is some subtly around how the J9UTF8 * are
compared as we can't do a string comparison and must rely on
utf8 interning as previous J9UTF8's may have been unloaded
between processing this frame and preceeding ones.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>