Only check the intern table when requesting interned strings
Previously, we would hash utf8 strings that didn't require translation
to see if we could reuse a string from the intern table.  This results
in a lot of utf8 hashing which isn't necessary.

This patch changes the algorithm to only probe if an interned string
is requested.  Otherwise, a new string will be allocated.

Also changed to use the VMConstantPool macros rather than looking up
the class long form.

Minor cleanups to privatize the J9InternalVMFunction table rather than
repeatedly fetching it off the VM.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>