Update IProfiler to handle invokePrivateMethod
IProfiler already was aware of the initialVirtualMethod due to
how the profiles are collected - the VM just writes the J9Method*
that corresponds to the vtable index without doing additional
processing.

Previously, only the initialVirtualMethod would show up in that
case.  Now the invokePrivateMethod can as well and will more
often with #7666, which stops doing the invokevirtual ->
invokespecial mapping for private methods.

Further improvements are needed to get profiling for the "real"
method being called in the invokeprivate case but that will
have to be done in a separate PR.

As part of this, I've also modified jitGetJ9MethodUsingIndex()
to consistently return NULL for the initial*Method and
invokePrivateMethod cases.  Though there is some very strange
code in one case that I believe was, and still is, incorrect.

fixes: #7696

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>