Call getROMString() to retrieve the signature at the server
When `setSignature()` is called at the JITServer in
`TR_ResolvedJ9JITServerMethod::getResolvedPossiblyPrivateVirtualMethod()`,
the signature might not have been stored in `romMethodRef` at the
server. `getROMString()` should be called to retrieve the signature
and it takes care of querying the client side if it hasn't been cached.
Otherwise, `jitParseSignature()` will be triggered with an empty
signature string. It doesn’t handle the case of signature length
being 0 and ends up going into an infinite loop and causing stack overflow.

Also made the change to call `handleUnresolvedSpecialMethodInCP`
instead of `handleUnresolvedVirtualMethodInCP` in
'TR_ResolvedJ9Method::getResolvedSpecialMethod()`.

Issue: #7567

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>