ACMP support for value types
This adds support for substitutability test in the acmp* bytecodes which
is needed to support value types.

Overview:
-----------

Instead of doing a reference comparison when the ifacmp* bytecodes are
executed with valueType operands, as is the case with reference type
today, valueTypes will perform a structural comparison instead
(substitutability test). This overload of the ifacmp* bytecodes is
required because valueTypes subclass java.lang.Object and use a*
bytecodes. It is possible for the static type of a valueType to be a
reference type (see example below) but the reflexive properties of the
`==` must still be maintained.

Object p = (Object) Point.makePoint(1, 2);
if (p == p) //must be true

The original acmp (legacy acmp) operation in the ifacmp* bytecodes
performs a simple ref comparison, lhs == rhs. This must be updated to
check if both operands are valueTypes, and  if they are of the same
type. If that is the case a structural comparison must be performed.
Example below:

```
acmpSub(lhs, rhs) {
        if (lhs == rhs) return true
        if ((lhs == NULL) || (rhs == NULL)) return false
        if ((lhs.class == rhs.class) && (lhs.isValue)) return
isSubstitutable(lhs, rhs)
        return false
}
```

The structural comparison must iterate through each field and perform an
equality comparison. If the field is byte, short, char, int the
comparison in icmp_eq is performed. If the field is a long the
comparison in lcmp is performed. If the field is a float an icmp
comparison of Float.floatToIntBits of both values is performed. If the
field is a double an lcmp comparison of Double.doubleToLongBits of both
values is performed. Note: this behaviour differs from the float f =
...; f == f case which performs a fcmp* operation. If the field is a
reference type that is java.lang.Object or an Interface or any type that
may contain a valueType (i.e InlineObject, Object, restricted-abstract
classes or interfaces) or if the field is a value type and both values
are not NULL a recursive acmpSub is performed. For all other types the
legacy acmp is performed. Example is below:

```
isSubstitutable(lhs, rhs) {
        for each primitive perform primitive comparison (icmp, lcmp,
Float.floatToIntBits then icmp, )
        for each ref that is j.l.Object, interface, a valueType and
both operands are not NULL perform acmSub (recursive)
        for all other refs perform legacy acmp
        return result
}
```

There are cases where the JVM may optimize the individual field
comparisons of the substitutability test and simply perform a memcmp
over the range of the operands. This optimization is possible in cases
where the substitutability test will not require recursion and there are
no floating point types.
The acmpSub will check to see if the fast comparison flag exists, if so
it will perform a memcmp with both operands that does something like
memcmp(lhs, rhs, lhs.size).

This is based on Adithya's work in
https://github.com/eclipse/openj9/pull/8133

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>