Assert J9Class for stacktrace can never be an arrayclass
Arrays don't define any of their own methods so they will
never be found as the defining class of a method based on
the bytecode pc.

This assert ensures that the algorithm used to create the
class name will be updated if arrays are ever changed to
allow their own methods.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>