Stop cache nestMembers[] on j.l.Class
The code originally cached the nestMembers[] on the Class
object to avoid calling the JNI native to fetch them.

The code should have been clone()ing the array before handing
it out to users otherwise they can modify it.  Rather than
adding the clone call here, I've removed the caching of the
array as there are no callers of nestMembers() in the JDK.

From an access perspective, the nestHost is the interesting
value and the one worth caching.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>