Cache fileNameString on j.l.Class object during stacktrace generation
When converting the array of saved PCs into a StackTraceElement[],
we spend a lot of time allocating filename strings.  The same string
may be allocated multiple times.

Previously, the code was optimized to avoid allocations when two
subsequent PCs were from the same file.

This change corrects some bugs in that code and further caches the
filename directly on the j.l.Class.  This catches both the case
where the class has already allocated a filename while still
reusing previously allocated filenames.

The later case is most relevant to multiple classes that have been
allocated in the same file.

related: #7776

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>