Add support for single level peeking in EA under HCR
Currently EA does not support peaking under HCR. This change supports adding peeking of
depth 1 under HCR by creating HCR guards which force heapficiation of potentially stack
allocated candidates ahead of peeked calls to ensure accidental escape does not happen
 at runtime even if peeking results are trusted.

The nature of this change relies on OSR to retrieve the live local information. Peeking
under HCR with OSR off is NOT supported. A protected call will be proceeded by an HCR
guard with heapificaitons on the taken side of the guard (if needed) to ensure that
stack allocated objects are heapified ahead of the call if the peeked method is
redefined and could potentially lead to an escape at runtime.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>