Fetch the module from the j.l.Class
Re-arrange the code so the J9Class is looked up earlier and
can be used for more operations.  In particular, use it
to optimzie the j.l.Class->j.l.Module->J9Module lookup to
avoid the packageNameLength calculation and locking aroound
the module-related hashtable.

Also, fix the potential null dereference of the classlaoader.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>