Improve write barrier codegen by eliminating redundant loads
When investigating TPCDS query 72, performance improved significantly when using the
optthroughput GC policy as opposed to gencon. Upon further examination, it was observed
that the space boundaries for the new/tenured spaces in the heap were being loaded from
memory twice: once when checking if the object is in the tenured space, and once when
checking if it is in the new space. This change is an attempt to save on performance
by keeping these values in reserved registers, thus eliminating the need for a second load.

Signed-off-by: Jackie Midroni <jackie.midroni@ibm.com>