Override needRelocationsForLookupEvaluationData
This commit overrides the needRelocationsForLookupEvaluationData
query found in the OMR codegenerator. The query is used to decide if
a reloction record is needed when generating code for the lookup IL in
power's code generator (it is needed for AOT and Remote compiles).
Since OMR is not aware of remote compilations, overriding the query here
will allow Openj9 to make a decision on whether a relocation record is
needed.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>