Deprecate j9gc_allocStringWithSharedCharData
The features this API supports are a strict subset of the features
supported by the `j9gc_createJavaLangString` API. There is currently
only one use of `j9gc_allocStringWithSharedCharData` which can be
changed to use the more complete `j9gc_createJavaLangString` API by
asking for the string to be tenured via the string flags. Note that the
resolve flags were never actually used within the API so that parameter
does not matter.

We take this opportunity to also clean up `getUTF8UnicodeLength` as we
no longer need to pass the string flags in since they were only used to
make the `j9gc_allocStringWithSharedCharData` API function correctly.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>